arm: add general MPU configuration options
This just adds the basic knobs and disables features on Kconfig level
that directly relate to an MMU.

We deliberately do not support running without any protection. That is,
if the CPU uses an MPU instead of an MMU the MPU is always used.
Likewise, there is no possibility to disable the MMU if the CPU has
one.

Change-Id: I3359495f32f0fc66ec659a0d5b6201e8a468c22b