arm: add TLB maintenance functions for MPU systems
In principle, no TLB maintenance is necessary on MPU based systems. On
arm32, branch predictor maintenance may still be required, though.
Also, if VMSA is supported at EL1, we again have to do TLB maintenance.
Quoting ARM DDI 0600B:

  Stage 1 VMSAv8-64 is permitted to cache stage 2 PMSAv8-64 MPU
  configuration as a part of the translation process. Visibility of
  stage 2 MPU updates for stage 1 VMSAv8-64 contexts requires
  associated TLB invalidation for stage 2.

We optimize to switch the VMID only if necessary. Still, the assumption
is that manipulating foreign tasks (i.e. ones that are not currently
active) is the likely case. Therefore there is no ISB after restoring
the original VMID because this is done implicitly when returning to
user space. This implies that we have to make an ISB before flushing
the TLB of the current task because the last VMID switch might not have
retired yet.

Change-Id: Id7f0a4294e2c0cd913f9958b83e72178d849741e