thread: Properly reset timeout in do_kill()
Resetting a timeout in do_kill() is only defensive programming anyway,
i.e. it should never happen because all callers of set_timeout() cleanup
via the IPC_timeout destructor.

But let's assume we really have a timeout set here, then the current
approach is insufficient, because it does not set the _timeout pointer
to nullptr. If a thread is resurrected later, the timeout (possibly
dangling pointer) might be re-enqueued again.

Change-Id: Id07c402765e084143a8a17f0b554e402a5bcef86