Restore correct guest value of RBX/EBX before VMLAUNCH
Since the change Icbb47a5f4c53b352a46b462ce5e0826be002be38, the RBX/EBX
register is used to read the VM-instruction error field after a failed
VMRESUME. In case the reason of the VMRESUME failure is a non-launched
VMCS, the VMLAUNCH instruction is used to launch the VMCS.

Unfortunately, the proper guest value of RBX/EBX was never restored
before launching the VMCS using VMLAUNCH.

This is not problematic if the VMLAUNCH instruction is used only for
the first initial launching of a static per-CPU VMCS (i.e. the current
implementation). The actual value of RBX/EBX hardly matters in this
case. Nevertheless, the code is still formally incorrect.

In the future, when we switch to a dynamic per-vCPU VMCS, a VMCS can
get into a non-launched state repeatedly while being deactivated and
therefore the guest context would be incorrect without restoring the
proper guest value of RBX/EBX.

Change-Id: I5d49f0e6e4fd1b97351c76f2602208fb35a932fa