kmem_alloc: hard code number of fixups
The fixups are global data that needs to be wrapped into a container
class in the future. Because the number of fixups then is a property of
the type, we cannot rely on a constexpr function anymore. The
underlying reason is that the parsing of function bodies is deferred
until after the class specifier. Hence it is impossible to use a
constexpr function in the same class definition for a type. Fortunately,
the overhead of the unused fixups is negligible.

Change-Id: I4ff68f0408b51875bdc0f9fd045a419beba06cb7