arm: add kernel entry/exit code for MPU systems
On kernel exit, the heap region must be constrained to the stack and
all user space ku_mem regions need to be enabled. Additionally the KIP
is made read-only. These steps need to be reverted on kernel entry.

As a performance optimization, the entry path does not execute an ISB
instruction after restoring the access to the kernel heap and the KIP.
Legally, this would be required because the MSR writes are only
guaranteed to be effective then. Practically, all Cortex-R processors
so far are in-order architectures. They would take a considerable
performance hit from the ISB while not needing it from a practical
point of view. Instead, a special in-kernel fault handler copes with
the potential fault that should never happen anyway.

Change-Id: I677362616a94981793555ea7f937c27083967733