arm: add MPU mem_space support
Implement the central functions to manipulate a Mem_space object. These
calls update the underlying Mpu_regions object and, if the Mem_space is
currently active, the changed regions are directly synchronized to the
MPU registers in hardware.

Speaking of TLB maintenance, if the map_util helpers resort to flushing
all TLBs, it implies to reload the currently active Mpu_regions into
the MPU. Because each CPU has it's own MPU registers, the same needs to
be done if a Mem_space is currently active on a remote CPU.

On context switch, also just load the new Mpu_regions into the MPU
registers. Additionally, the mask of ku_mem regions must also be
synchronized to the current context so that the kernel exit code can
enable the respective regions.

Change-Id: I145f1d113756fca5418e37c94168faffe1f30882