arm: update linker script for no-MMU systems
Because there is no virtual memory without an MMU, all addresses are
physical addresses. To keep things simple, the whole kernel is still
linked position dependent.

The start of the kernel includes the KIP on MMU sytems while the KIP is
excluded on no-MMU systems. This is caused by the fact that the KIP is
mapped by a dedicated MPU region and it is not allowed to have
overlapping regions.

Change-Id: Ia23e12a53cf60d4c81f7c93d4d7292f611ce96de