arm: add MPU related state to context
On Armv8-R MPU regions must not overlap. This is an issue for the
kernel because ku_mem regions are allocated from the kernel heap. The
kernel entry and exit paths must thus enable such regions on kernel
exit and disable them on re-entry. Additionally the kernel heap region
must be constrained to the stack, otherwise the user space could still
not access the ku_mem because of the overlap.

To make the kernel entry and exit as lightweight as possible, all
required information is stored in the currently active context. This
needs to be implemented in Context_base because the Mem_space code also
needs to update these fields.

Change-Id: Ide548f942cca27d5013025c1c9a2730ad369a41f