arm: setup exception entry points on MPU systems
On arm32 without virtualization, the exception vector table can be part
of the kernel image, like it is already on arm64. It just needs to be
aligned sufficiently.

The initialization for virtualization enabled kernels is very similar
to MMU kernels. On arm64 most code can be shared. The major difference
is the V(S)TCR register that is secure on MPU kernels and has some
additional bits.

Change-Id: I49c597c5b7c65cde3d0c566b5598db8e89919765