arm: refactor generic timer matching
Each platform potentially has a different mapping of generic timer
PPIs. This mapping is done in the BSP and, depending on the kernel
configuration, the right interrupt number is chosen.

Depending on the platform, not all architecturally defined timers are
actually implemented. This makes the matching through an enum more
complicated because the compiler will warn if some enum values are not
used in the switch/case. Instead, use an integer constant to avoid the
potential warnings.

Change-Id: I066a089674cc75ad04dbe0bacb4c43013b09f34a