global_data: add wrapper for global data
In case of an AMP setup, the kernel image will be reused across
multiple nodes (read: non cache coherent CPUs) to save memory
resources. This requires that there is no global shared data between
the nodes, though.

The Global_data template class is used to wrap such global data to
accomplish two things: put the data into a dedicated section and to
redirect accesses to the node-local backing storage. This obviously
needs to be done only for the AMP-case. In the regular, non-AMP-case
these classes try to be as transparent as possible.

In a way this is roughly equivalent to thread-local data.
Unfortunately, at least for gcc, it is not possible to combine
thread-local data and initialization priorities. Many parts to Fiasco
rely on initialization priorities to steer the construction order of
global objects, making the thread-local data approach a non-starter.

Change-Id: If74c257d288ca02eb5988219e84c115a91defb17