static_init: add constinit init priority
With the introduction of the Global_data wrapper, all wrapped global
variables are potentially dynamically initialized. This can mess up the
initialization order of objects that are "zero- or
constant-initialized". These objects are dynamically initialized in the
AMP case.

Add an initialization priority for such objects so that these variables
can still be initialized first. To make this even more expressive,
DEFINE_GLOBAL_CONSTINIT should be used where applicable. The macro will
expand to 'constinit' once we enable C++20 in regular builds and
transparently use the right initialization priority on AMP builds.

Change-Id: I63215e029bdc309184d19f191ad6b193f6813b2f