kmem: avoid reinterpret_cast for initial kpdir
The storage to the initial kernel page directory is allocated in the
bss section. Its pointer is forwarded to the bootstrap stage as well as
being used to initialize Kmem::kdir.

Later, Kmem:kpdir will be wrapped into a Global_data. To retain the
constant initialization property, the reinterpret_cast must be avoided
because the language forbids constexpr expressions with
reinterpret_cast. Instead, use a union to achieve the same result but
avoid any cast.

Change-Id: I1f30cc8aab3fdf8c809c0b5775785bc52e36f9ba