thread: Prevent enqueue of new IPC senders for dying thread
The fine-grained cpu lock management in Thread::do_kill() suggests that
the function was written with the assumption that the cpu lock is not
necessarily held when calling the function.

However, if the function in its current implementation would really be
called without the cpu lock held, the thread can be preempted after the
sender dequeue loop that aborts all pending IPC senders.

Being preemptible after the dequeue loop and before setting the
Thread_dead flag means that new senders can be enqueued in the sender
list of the to be killed thread, because Thread::check_sender() only
checks for Thread_dead, not for Thread_dying.

Fortunately, however, it seems like Thread::do_kill() is only ever
invoked with the cpu lock held.

Change-Id: I513a4646e13786bca669945bc574dea30d2c65e6