thread: Simplify interrupt masking in Thread::do_kill()
Contrary to what the fine-grained CPU lock management in
Thread::do_kill() suggests, the function seems to be always invoked
with interrupts disabled (CPU lock held), since it is executed as a
continuation, where interrupts should already be disabled.

So in theory an assert that the CPU lock is held should be sufficient,
but it does not hurt to really make sure by grabbing the CPU lock at
the start of the function.

Change-Id: I8f3a072872631798006b943109c46c7a66110d7e