abi: optionally select destination task of receive buffer
By default, the task of the receiving thread is the destination of any
transferred capabilities. Sometimes it is useful that the receiving
task is different, though. This is usually the case when some mapping
from task A to task B shall be established by some third party, living
in an unrelated task C. Traditionally this required to first make a
mapping from A to C and then from C to B.

This overhead can be avoided by (optionally) passing the destination
task explicitly. If the compound bit of a receive buffer item is set,
an additional word with the capability index of the destination task is
passed. This capability slot is looked up in the receivers object space
to fetch the destination task. The receiver naturally needs to have
write rights for the destination task.

See ticket #L4RE-300.

Change-Id: I64bd57a6125f412d82db31f341f7c5ca6d9aec3c