io_space: Simplify IO bitmap allocation
Each space has its own IO bitmap, which the space itself allocates and
delets. In particular, the bitmap is not created by a space, then its
ownership transferred to another space, and then later deleted by that
other space. It is therefore not necessary to record the quota used for
allocation in the bitmap, as the quota of the space is used for both
alloc and delete.

Change-Id: I21a1d0aa75f8099c040cd3b7ccde796ea89d6e1a