Fix deferred deletion of thread
initiate_deletion() is invoked when the last capability to the Thread is
released. Then after invalidating the existence lock and waiting for an
RCU grace period, the "capability" reference (corresponds to the
inc_ref() in the constructor of Thread) is released.

If someone else holds an additional kernel-internal reference to the
thread, deletion of the thread is deferred until that someone releases
its reference. For thread that is done via Thread::put_n_reap(), which
then surprisingly starts the initiate_deletion() procedure again.

This seems unnecessary or potentially even broken. For the other
reference counted object Task it is not done, the task is just deleted.
The same should be possible for Thread. In fact, it used to be handled
this way until commit 8e1d9801a05de1ebcd3d2f9bcf80e1e0f14a443c changed
it. But its description indicates that the change was not a bug fix, but
just a refactoring.

In addition, thread currently cannot be used with Ref_ptr because it
does not know about the speciality of Thread and thus calls delete
directly if the reference count drops to zero.

So remove Thread::put_n_reap() and instead invoke delete directly.

Change-Id: I55b0f439968ccd40b54c2d800f2daf31f2955f94