Del_irq_chip::unbind() shall call Irq_chip_soft::unbind()
It is not sufficient to remove the reference from the thread to the
deletion IRQ but we also need to remove the reference from the deletion
IRQ to the thread.

The latter is performed by explicitly calling Irq_chip_soft::unbind()
which binds the deletion IRQ to the SW chip. As a result, the chip of
the deletion IRQ no longer points to the Del_irq_chip member of the
thread.

See ticket #CD-1168.

Change-Id: I9a10f515e7b545c98bd16a61bd43ae4197dd4370