Arm: Do not allow sync kip clock for A5+A8+A9
The Cortex-A9, Cortex-A8 and Cortex-A5 do not use the generic timer but the MP
timer and thus cannot use the sync-kip clock feature, as there's no good
counter set up and user-level has no means to read the timer (reasonable).
Depend on the virtualization features as this is the difference.

Change-Id: Ib82a175f94bf83d9d04a65280269641c28fe3c4d