preprocess: Add support for constexpr
Annotating a function with constexpr implies inline and in addition
functions annotated with constexpr must always be generated in-line.
Preprocess however did not know about constexpr and therefore did not
handle constexpr functions correctly:
  - constexpr
      -> Generated in-line.
  - inline constexpr:
      -> Generated in-line, but not always, only if -i option is used.
  - inline constexpr ALWAYS_INLINE:
      -> Always generated in-line, but forces the compiler to always
         inline the function, this might not always be desired. In
	 addition it is cumbersome to type.

Preprocess now recognizes constexpr functions (no inline specifier
needed) and treats them as inline function, with the only difference
that they always generated in-line even if the -i option is not used.

Change-Id: I2930d9d188814dcbb38d11cb4ffa7b5a0457f478
Requires-Change-Id: I35984f12000d8a1058301e6364b5fd71df5eec25