irq_sender: Split re-configuration in locked and unlocked phase
The previous approach of holding a lock over the entire re-configuration
of the Irq_sender is unfortunately incorrect, as revoking the IRQ from
the old target thread potentially requires executing a cross-core DRQ.
That means the thread doing the re-configuration might block (waiting
for a DRQ reply) which is not allowed/possible while holding either a
Switch_lock or Spin_lock.

Therefore, split the re-configuration of the Irq_sender into two phases:
The first phase, running under the Irq_sender's _irq_lock spin lock,
exchanges the old target thread with the new target.
The second phase, not running under any lock, revokes the IRQ from the
old target thread and then re-injects the IRQ to the new thread if it
was pending.

See ticket #CD-1309.

Change-Id: I8e7766ed1d6fd2d1abf5a27c4d6e2f42af8e86e4
Requires-Change-Id: Ic940faec375713635c39e7379b613b9bd36af247