irq_sender: Remove from DRQ queue and IPC sender list on delete
While detaching the thread, an IRQ hit might resulted in Irq_sender
being queued in an IPC send to a thread. Hence, the Irq_sender cannot
be deleted directly, but only after this IPC send has been completed.

This is coordinated by taking a counted reference in
Irq_sender::destroy(), which is released by whomever transitions the
Irq_sender from Invalidated to Destroyed state. This can be the thread
that initiated the deletion of the Irq_sender or the receiver thread on
finishing the IPC send.

See ticket #CD-759.

Change-Id: I53f710ddd62b64ec44cfa5653f4ce0673ec10025