ipc_sender: Dequeue IRQ sender if receiver aborts IPC
When the receiver aborts an IPC, only happens when a receiver thread is
killed, it informs the sender by calling ipc_receiver_aborted().

But the implementation of ipc_receiver_aborted() for Irq_sender missed
to consume the IRQ. Therefore, the Irq_sener still assumed that an IRQ
send operation was already pending and consequently did not send a new
one, even if it was bound to a new thread.

See ticket #CD-759.

Change-Id: Ie54c710e2ef8ca39579a0c6344f41937c84766db