cpu_call: Make async calls more visible and enforce correct func type
The async parameter was incorrectly documented, an async cpu_call_many
does not wait until the function is executed on all target CPUs. It only
waits until the func is dequeued from the Cpu_call_queue on all target
CPUs.

Like stated in the original commit message that introduced the async
feature, therefore an asynchronously executed func must not be a
capturing lambda, as the func might execute on a target CPU long
after the caller has returned and moved on (see also the documentation
of cxx::functor).

Also adjust the callers of cpu_call_many.

See ticket #CD-1244.

Change-Id: Ie7817760fcda12283db5c134e5c3accd8be64f6c