cpu_call: Fix two race conditions with async cpu_call
An asynchronous Cpu_call was considered done, once the Cpu_call was
dequeued from the Cpu_call_queue of its target CPU (Cpu_call::queued()
returns false).

When a Cpu_call is considered done, it can be used for a call to another
CPU if necessary. Or, if all Cpu_calls are done, the cpu_call_many
function returns, in which case the Cpu_call object is freed (allocated
as part of an array on the stack).

This resulted in two problems:
1. An async Cpu_call is considered done (might be freed) before the
   remote CPU is done with it, i.e. read the function pointer from the
   Cpu_call. In the meantime the caller might have put some other value
   at that location on its stack.

2. For a sync Cpu_call a write to the Cpu_call::_wait member is used to
   signal that function execution on the remote CPU is done.
   Unfortunately that write was also done after executing an
   asynchronous Cpu_call, potentially on an already freed Cpu_call.
   That means at a "random" point in time the remote CPU zeroes a
   "random" Mword on the callers stack.

To solve these two problems, asynchronous Cpu_calls now also use the
Cpu_call::_wait member. However, they set it before calling the func.

See ticket #CD-1244.

Change-Id: If96963802692f3baa752b157cfcdbf524239db10