cpu_call: Require also for async calls that CPU lock is not held
The difference between asynchronous and synchronous execution is that
for the former the call is marked as done immediately before executing
the function whereas for the latter the call is marked as done after
executing the function. In both cases, however, the cpu calls are first
queued into the cpu call queue of the target CPU until it processes its
queue.

Therefore both sync and async calls can result in a deadlock:
1. CPU A queues async or sync cpu call at CPU B.
2. CPU B queues async or sync cpu call at CPU A.
3. Both CPU A and CPU B wait with interrupts closed for the other CPU
   to process its cpu call queue.

See ticket #CD-1244.

Change-Id: Iceb575c5327167623480b4d460cc9ef58315d028
Requires-Change-Id: If3ee610fa70c1ac0a672f1608d156e7d4d93c509
Requires-Change-Id: Iaa7a044890678c3c817b0e287f982e2d99b344ef