ARM/switch_vm_state: save/load EL1 state only if necessary
Save the EL1 state only if switching from a privileged thread.
Load the EL1 state only if switching to a privileged thread.
Load default values if not switching to a privileged thread when
switching away from a privileged thread, when enabling extended
vCPU state, or if a thread enables EL1.
HCR_EL2 and CPACR_EL1 must always be saved/restored.

Change-Id: Id4568207a1a2d2f1545314087a31fc8c78a0051c