Remove confusing comments
Without an explanation (which is not even present in the original
commit message), the remarks about locking the current task being
unnecessary is confusing at best.

The basic understanding of the code actually suggests that locking
the current task is essential for the mutual exclusion in the mapping
and unmapping routines.

Change-Id: Ia83231367bb0e4c987b75a31fdde38501e43c543