arm64: Use PXN together with UXN on !cpu_virt
In !cpu_virt configuration, the functionality of the execute-never bit
is split between UXN and PXN bits, where UXN overlaps with the bit
definition of XN in the cpu_virt configuration. The previously unused
PXN bit is essential for the future implementation of non-executable
kernel mappings in !cpu_virt kernels.

This commit conflates the meaning of UXN and PXN bits when dealing with
page tables so that they are set together. An exception to that are
kernel PTEs that always have UXN set independent of PXN.

At the same time, this commit makes sure the RES1 bit AP[1] field is
always set in cpu_virt configuration.

Fixes issue #CD-1255.
Fixes issue #CD-1257.

Requires-Change-Id: If05f82f6dbff92b3fa9dbbd701afc6355adbf480
Change-Id: I3728f35d98641e2c82bb54958b27f2a393871f7d