VMX: Better naming of a class and a template parameter
Rename the Vm_vmx_b class to Vm_vmx_base to capture that it is a base
class of the two VMX implementation variants (without EPT support and
with EPT support). Rename the template parameter X to VARIANT to
capture that it discriminates the two implementation variants.

No change in functionality.

Change-Id: Ifab143b800a3909addaee03f7533a5b86f5d8b75