jdb/arm: disable RX IRQ on JDB entry, re-enable on JDB exit
This is a workaround for the Linflex UART: In FIFO mode, the RX IRQ is
triggered when the RX queue gets empty, not when more than one character
is received. Therefore we have to disable FIFO as long as we depend on
the RX interrupt -- outside JDB -- but enable FIFO mode after entering
JDB.

As of Ia93d71e9cfe5d565a7e48932dc6e765e84d1e087, disabling the RX IRQ
has the side effect of enabling RX FIFO mode for the Linflex UART, and
enabling the RX IRQ also disables RX FIFO mode for the Linflex UART.

Disabling/Re-enabling the RX IRQ on entering/leaving JDB should not
cause any harm on other UARTs. For now we do this only on ARM.

See ticket #CD-964.

Change-Id: Idd47fc483d270c9022025611e146feffe48fed3f