[API] l4sys: fix potential register corruption on arm64
When linking shared, the dynamic linker will resolve symbols lazily at
runtime. The resolver will obey the calling convention and thus safe
only arguments and callee saved registers. Because __l4_sys_syscall is
such a global symbol, we must force the caller to spill the caller
saved registers.

The solution is to make __l4_sys_syscall compliant to the standard ABI
on arm64. This way we can declare it as an extern-C function and let
the compiler take care of all the dirty details.

Change-Id: I26e1d5eef3d3a140464d78ccbdfbdcafd1fe3fcb