thread: do not trap alien threads on fpu ctx alloc
The observable behaviour of the alien thread by the exception handler
depends on the kernel configuration. The trap can only be observed on
lazy_fpu configurations. Additionally, there is nothing the exception
handler can do to resume execution of the alien thread in this case.

In this light, let's remove the trap for the time being. It can be
easily added again when a proper interface has been implemented to
enable/disable FPU access of an (alien) thread. So far, no user space
code seems to rely on the feature anyway.

Change-Id: Ic24ec14eb388362b2f1c71af7b99fdf55e7416fa