Enable compilation using LLVM toolchain
llvm-objcopy does not support --adjust-section-vma option which was used
by the build system in a somewhat hacky way as a primitive linker. Since
the result is precisely the two sections in their binary form
concatenated, --adjust-section-vma can be avoided by converting the two
sections to binary separately and then concatenating them.

This change also makes the code more robust because the offset is no
longer hardcoded.

"Contribution covered by the CLA"

Change-Id: I00f06759c31d50fca5de4c536ae343164b9fc66b