introduce INIT_WORKLOAD{,_APP_CORE} mechanism and convert std workload
Use .init_workload.* sections to store function pointers to initialize
the kernel workload for the boot CPU and .init_workload_app_core.* to
initialize the kernel workload for the application CPUs.

The default workload for the boot CPU is now initialized in the static
function init_std_workload() which is only added to regular kernel
images. Kernel unit tests initialize the workload by init_unittest()
instead. This is no longer a weak symbol.

Normally there is no default workload for application CPUs. Some kernel
unit tests use INIT_WORKLOAD_APP_CORE() announcing a function pointer
which is called by an application CPU before it enters the idle loop.

The benefit of this change is more flexibility: It is possible to add
more function pointers to perform other jobs before entering the idle
loop. And the function to initialize the standard workload is no longer
linked to kernel unit tests.

Change-Id: I4f760b1d99b210fe9b6bdb6d998fb396fb2316d9