arm: fix 32 bit !cpu_virt execution
The definition of the LPAE page tables differs slightly between 32- and
64-bit execution. In both cases, bit 53 is defined as PXN (Privileged
Execute Never). But on AArch32 bit 54 is defined as XN (Execute Never)
which applies to all exception levels whereas AArch64 defines bit 54 as
UXN (Unprivileged Execute Never) which only applies to EL0.

To make the distinction manageable, define all three constants: PXN,
UXN and XN. Depending on the execution mode, some of the bits are not
available. This is fine because the PTE code tries to use all available
features and unavailable bits are optimized away by the compiler.

Change-Id: I4fb595a611dffd7bf7675be320fd37545e4dc7e6