arm: fix execution on non-standard AMP nodes
Even if fiasco is built without AMP support, the platform might have a
notion of AMP nodes. Specifically, if fiasco is built for the 2nd
cluster on the S32Z, it's single core will have the AMP node ID 4.
Bootstrap expects this very number as node-ID in the KIP. This implies
that even single core builds should use the right AMP node-ID in the
KIP.

Instead of using a plain constant, use a static constexpr inline
function to provide the compile time constant. The default
implementation that starts at 0 can be overridden by the platform as
necessary. This way, most platforms do not have to deal with the
first_node() and the code gets independent of the "amp" preprocess
condition.

Change-Id: I9b7c8fa6ac41d515a6c2f1f986a0ab61ed67750f