task/map: destroy the reap list after destroying the lock guard
A thread must not destroy any kernel object where this thread has the
existence lock acquired for. To fix an edge case where the target for
the mapping operation is added to the reap list by map_util/map(),
ensure that the reap list is created before the lock guard.

See ticket #CD-1280.

Change-Id: I319e7b5d0eb081c8bf47807d1b1b6b9c2227bd59