ia32: introduce TSC_UNIFIED
If this option is enabled then the kernel assumes that the TSC is
synchronized across all CPU cores. For now, this requires the
invariant+unified TSC feature assuming that the cores are already
synchronized and don't stop during any ACPI state.

With CPU synchronization during boot it would be also sufficient to
depend on no-stopping TSCs running at a constant rate on every CPU.

Change-Id: I579c49b0aedd10510075545c2b08938c89941020