cxx: Let functor copy for function pointers
The cxx::functor stores only an address of the Functor object passed to
its constructor, i.e. it does not copy the object itself. This means the
caller must be careful not to pass a Functor object that is out-lived by
the functor instance.

This change introduces an exception for function pointers. They are now
copied, and thus are not affected by the above restriction.

For Functor objects, such as capturing and non-capturing lambda
expressions, the behavior of cxx::functor remains unchanged.

See ticket #CD-1244.

Change-Id: I3c113dcc5c983bbfa0eb2d1409385bc7af82519b
Requires-Change-Id: I3fdfd36d64056f442d18e882e005d8e8ba955502
Requires-Change-Id: Ib8b7121491834f11bf7a73422401158e79f4a4f9