arm: do not read ID_DFR1 register on MPU systems
Officially, the ID_DFR1 register is "unallocated" on Armv8-R AArch32.
It's readable on the Cortex-R52 and Arm FVP, though. Additionally, it's
presence is architecturally implied because a PMUv3 (which is supported
by the R52) may have the multi-threaded PMU extensions and the ID_DFR1
register is the canonical way to discover this feature.

Change-Id: I7da90944cb70ed99b7c6ddc873a31b9315ac0454