thread-arm-hyp: consider that a hyp mode trap could be handled
If we enter the nested trap handler for handling a hyp mode exception,
consider that the handler was able to deal with the exception and
return normally. This is usually the case with JDB.

This is important for handling exceptions during system register access
from Jdb_kern_info_cpu: This change repairs 'kc' on ARM32.

Change-Id: I005228b517e9e9665ce6320090961b084c96db9b