per_cpu_data: fix Per_cpu_ptr with Static_object
Internally, the Per_cpu objects use Global_data members. This class has
a specialization for Static_object to make the Global_data wrapping as
transparent as possible. But this overload lacks the address-of
operator. Hence we have to unwrap the Global_data object to get a
consistent behaviour, regardless of the underlying type.

Change-Id: I5c0ee70ef13a2b9c282df5a0a2d388e4686c4b89