types: de-obfuscate nonull_static_cast
A static_cast is not just a simple pointer adjustment. If given a
nullptr, the nullptr must be retained. The nonull_static_cast template
is used to inform the compiler that we know the pointer is not null and
the nullptr handling can be omitted.

Nowadays gcc and clang support the __builtin_unreachable() builtin to
communicate impossible conditions to the complier. Using it, both
compilers are able to elide the nullptr handling of static_cast.

Change-Id: Ied0e0081411487c27e08c2ef41dcdd210229515d