timer_tick: never call handler_all() with nullptr
Some platforms use a hack to call Timer_tick::handler_all() with a
nullptr in the knowledge that the passed pointer is practically unused.
This is still wrong because the generic Timer_tick code uses
nonull_static_cast() which directly contradicts the caller.

Rectify this by introducing a static handler_all_noack() function that
does not require an Irq_base object to be passed. It's the duty of the
caller to acknowledge the interrupt in this case.

Change-Id: Id944074971189e02cdc64fb07df811445b94a398