introduce Timer::aux_clock_unstopped()
In certain places we need to read a clock but we run on the JDB stack
where current_cpu() does not work and where we do not want to update the
KIP clock as side effect for reading the system clock. In these cases it
is not relevant if the clock is synchronized with userland but the clock
must run continuously. Introduce Timer::aux_clock_unstopped() for this
purpose. Note that this clock might be only available with certain
configurations.

The existing function Timer::system_clock() shall be used if the clock
value is related to userland: If the KIP clock is synchronized with the
timer the kernel is using as internal clock, the KIP clock should be
updated more often (by threads running on the boot CPU) so that userland
and kernel have a consistent view on the system time.

This is the first step for making a direct access to Kip::k()->clock()
from the kernel superfluous.

Also unify the variable names where Time::system_clock() is used.

Change-Id: Id02331b9251bc7e627eb3da064e9397a78897205